Work in progress: update to Jackson 2.9.9
This is a work in progress.

Behavior changed upstream in Jackson 2.9's MapSerializer to pass entire
map object to the filter instead of the entry.getValue() for the
specific entry.getKey() being filtered. This means that the object
received by the filter is of an unexpected Map type instead of the
Collection (in the JsonOutputFilter, the value is a Collection, because
it's the set of values coming from a Multimap).

This change occurred to address FasterXML/jackson-databind#1655 and
occurred in commit a2c667b9b8482581e64bb8752963984efb1214b5
(https://github.com/FasterXML/jackson-databind/commit/a2c667b9b8482581e64bb8752963984efb1214b5)

Another change seems to have affected the filtered output. The output
that is emitted is not the filtered output, but the unfiltered output,
as if the filter is running, but its output does not affect
serialization.

This is a very strange and buggy result.

Earlier versions of Jackson databind are affected by numerous CVEs, but
I'm not sure how to work around this (possibly broken) behavior change
in order to upgrade to 2.9.9

This commit demonstrates a partial workaround, fixing the type change...
but I have not yet figured out how to fix the second issue to have the
filter affect the serialized output properly, and this is demonstrated
by a test case which continues to fail.